#37 Rename: post image에서 post file로 이름 변경
이미지 및 영상 같이 저장할 예정임으로 파일로 이름 수정
