+++ src/main/java/com/codiary/backend/global/service/PostFileService/PostFileCommandServiceImpl.java
@@ -1,4 +1,4 @@
-package com.codiary.backend.global.service.PostPhotoService;
+package com.codiary.backend.global.service.PostFileService;
 
-public class PostPhotoCommandServiceImpl {
+public class PostFileCommandServiceImpl {
 }
